clean code

+++ app/src/main/java/org/jaudiotagger/audio/mp3/MP3AudioHeader.java
@@ -585,7 +585,7 @@ public class MP3AudioHeader implements AudioHeader
         {
             if (mp3XingFrame.isAudioSizeEnabled() && mp3XingFrame.getAudioSize() > 0)
             {
-                bitrate = (long) ((mp3XingFrame.getAudioSize() * CONVERTS_BYTE_TO_BITS) / (timePerFrame * getNumberOfFrames() * CONVERT_TO_KILOBITS));
+                bitrate = (long) (((long) mp3XingFrame.getAudioSize() * CONVERTS_BYTE_TO_BITS) / (timePerFrame * getNumberOfFrames() * CONVERT_TO_KILOBITS));
             }
             else
             {
@@ -596,7 +596,7 @@ public class MP3AudioHeader implements AudioHeader
         {
             if (mp3VbriFrame.getAudioSize() > 0)
             {
-                bitrate = (long) ((mp3VbriFrame.getAudioSize() * CONVERTS_BYTE_TO_BITS) / (timePerFrame * getNumberOfFrames() * CONVERT_TO_KILOBITS));
+                bitrate = (long) (((long) mp3VbriFrame.getAudioSize() * CONVERTS_BYTE_TO_BITS) / (timePerFrame * getNumberOfFrames() * CONVERT_TO_KILOBITS));
             }
             else
             {
